+++ src/Ambev.DeveloperEvaluation.Common/Pagination/PaginatedResponse.cs
@@ -1,0 +1,11 @@
+﻿using Ambev.DeveloperEvaluation.Common.Response;
+
+namespace Ambev.DeveloperEvaluation.Common.Pagination
+{
+    public class PaginatedResponse<T> : ApiResponseWithData<IEnumerable<T>>
+    {
+        public int CurrentPage { get; set; }
+        public int TotalPages { get; set; }
+        public int TotalCount { get; set; }
+    }
+}
