+++ src/Ambev.DeveloperEvaluation.Common/Response/ApiResponse.cs
@@ -1,0 +1,12 @@
+﻿using Ambev.DeveloperEvaluation.Common.Validation;
+
+namespace Ambev.DeveloperEvaluation.Common.Response
+{
+    public class ApiResponse
+    {
+        public bool Success { get; set; }
+        public string Message { get; set; } = string.Empty;
+        public IEnumerable<ValidationErrorDetail> Errors { get; set; } = [];
+    }
+
+}
