+++ src/Ambev.DeveloperEvaluation.Common/Response/ApiResponseWithData.cs
@@ -1,0 +1,8 @@
+﻿namespace Ambev.DeveloperEvaluation.Common.Response
+{
+    public class ApiResponseWithData<T> : ApiResponse
+    {
+        public T? Data { get; set; }
+    }
+
+}
