--- src/Ambev.DeveloperEvaluation.WebApi/Common/ApiResponse.cs
@@ -1,10 +1,0 @@
-﻿using Ambev.DeveloperEvaluation.Common.Validation;
-
-namespace Ambev.DeveloperEvaluation.WebApi.Common;
-
-public class ApiResponse
-{
-    public bool Success { get; set; }
-    public string Message { get; set; } = string.Empty;
-    public IEnumerable<ValidationErrorDetail> Errors { get; set; } = [];
-}
