--- src/Ambev.DeveloperEvaluation.WebApi/Common/ApiResponseWithData.cs
@@ -1,6 +1,0 @@
-﻿namespace Ambev.DeveloperEvaluation.WebApi.Common;
-
-public class ApiResponseWithData<T> : ApiResponse
-{
-    public T? Data { get; set; }
-}
