--- src/Ambev.DeveloperEvaluation.WebApi/Common/PaginatedList.cs
@@ -1,31 +1,0 @@
-﻿using Microsoft.EntityFrameworkCore;
-
-namespace Ambev.DeveloperEvaluation.WebApi.Common;
-
-public class PaginatedList<T> : List<T>
-{
-    public int CurrentPage { get; private set; }
-    public int TotalPages { get; private set; }
-    public int PageSize { get; private set; }
-    public int TotalCount { get; private set; }
-
-    public bool HasPrevious => CurrentPage > 1;
-    public bool HasNext => CurrentPage < TotalPages;
-
-    public PaginatedList(List<T> items, int count, int pageNumber, int pageSize)
-    {
-        TotalCount = count;
-        PageSize = pageSize;
-        CurrentPage = pageNumber;
-        TotalPages = (int)Math.Ceiling(count / (double)pageSize);
-
-        AddRange(items);
-    }
-
-    public static async Task<PaginatedList<T>> CreateAsync(IQueryable<T> source, int pageNumber, int pageSize)
-    {
-        var count = await source.CountAsync();
-        var items = await source.Skip((pageNumber - 1) * pageSize).Take(pageSize).ToListAsync();
-        return new PaginatedList<T>(items, count, pageNumber, pageSize);
-    }
-}
