--- src/Ambev.DeveloperEvaluation.WebApi/Common/PaginatedResponse.cs
@@ -1,8 +1,0 @@
-﻿namespace Ambev.DeveloperEvaluation.WebApi.Common;
-
-public class PaginatedResponse<T> : ApiResponseWithData<IEnumerable<T>>
-{
-    public int CurrentPage { get; set; }
-    public int TotalPages { get; set; }
-    public int TotalCount { get; set; }
-}
