+++ src/Ambev.DeveloperEvaluation.WebApi/Features/Auth/AuthController.cs
@@ -4,6 +4,7 @@ using AutoMapper;
 using Ambev.DeveloperEvaluation.WebApi.Common;
 using Ambev.DeveloperEvaluation.WebApi.Features.Auth.AuthenticateUserFeature;
 using Ambev.DeveloperEvaluation.Application.Auth.AuthenticateUser;
+using Ambev.DeveloperEvaluation.Common.Response;
 
 namespace Ambev.DeveloperEvaluation.WebApi.Features.Auth;
