+++ src/Ambev.DeveloperEvaluation.WebApi/Features/Users/UsersController.cs
@@ -8,6 +8,7 @@ using Ambev.DeveloperEvaluation.WebApi.Features.Users.DeleteUser;
 using Ambev.DeveloperEvaluation.Application.Users.CreateUser;
 using Ambev.DeveloperEvaluation.Application.Users.GetUser;
 using Ambev.DeveloperEvaluation.Application.Users.DeleteUser;
+using Ambev.DeveloperEvaluation.Common.Response;
 
 namespace Ambev.DeveloperEvaluation.WebApi.Features.Users;
